% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{model_genes}
\alias{model_genes}
\title{Read the model's required gene set}
\usage{
model_genes(model)
}
\arguments{
\item{model}{A model loaded by \code{\link[=load_model]{load_model()}}.}
}
\value{
Character vector of gene IDs the model requires.
}
\description{
Read the model's required gene set
}
