% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{model_normalisation}
\alias{model_normalisation}
\title{Read the model's normalisation method}
\usage{
model_normalisation(model)
}
\arguments{
\item{model}{A model loaded by \code{\link[=load_model]{load_model()}}.}
}
\value{
Character scalar, one of the TimeTeller normalisation methods.
}
\description{
Read the model's normalisation method
}
