% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_test_projection}
\alias{plot_test_projection}
\title{Plot the training projection with the test data overlaid, in 3D}
\usage{
plot_test_projection(object, selected_local_projection, density = FALSE, ...)
}
\arguments{
\item{object}{A post-projection object from \code{\link[=project_test_data]{project_test_data()}}.}

\item{selected_local_projection}{One of \code{\link[=local_projection_choices]{local_projection_choices()}}.}

\item{density}{Draw covariance ellipsoids? Kept \code{FALSE} by default (see
\code{\link[=plot_training_projection]{plot_training_projection()}}).}

\item{...}{Forwarded to \code{plot_3d_projection_with_test()}.}
}
\value{
A plotly object.
}
\description{
Wraps TimeTeller's \code{plot_3d_projection_with_test()}: the test samples
projected on top of the training projection. Requires a completed projection
(the test data must already be present in \code{object}). Returns a plotly object.
}
